Refactor ChatMessage component to improve performance
